+++ graphistry/PlotterBase.py
@@ -4,7 +4,7 @@ import copy, hashlib, logging, numpy as np, pandas as pd, pyarrow as pa, sys, uu
 from functools import lru_cache
 from weakref import WeakValueDictionary
 
-from .util import (error, in_ipython, make_iframe, random_string, warn)
+from .util import (error, in_ipython, in_databricks, make_iframe, random_string, warn)
 
 from .bolt_util import (
     bolt_graph_to_edges_dataframe,
@@ -36,6 +36,13 @@ maybe_dask_cudf = None
 try:
     import dask_cudf
     maybe_dask_cudf = dask_cudf
+except ImportError:
+    1
+
+maybe_spark = None
+try:
+    import pyspark
+    maybe_spark = pyspark
 except ImportError:
     1
 
@@ -1135,7 +1142,10 @@ class PlotterBase(Plottable):
         return res
 
 
-    def plot(self, graph=None, nodes=None, name=None, description=None, render=None, skip_upload=False, as_files=False, memoize=True):  # noqa: C901
+    def plot(
+        self, graph=None, nodes=None, name=None, description=None, render=None, skip_upload=False, as_files=False, memoize=True,
+        extra_html="", override_html_style=None
+    ):  # noqa: C901
         """Upload data to the Graphistry server and show as an iframe of it.
 
         Uses the currently bound schema structure and visual encodings.
@@ -1166,6 +1176,12 @@ class PlotterBase(Plottable):
 
         :param memoize: Tries to memoize pandas/cudf->arrow conversion, including skipping upload. Default on.
         :type memoize: bool
+
+        :param extra_html: Allow injecting arbitrary HTML into the visualization iframe.
+        :type extra_html: Optional[str]
+
+        :param override_html_style: Set fully custom style tag.
+        :type override_html_style: Optional[str]
 
         **Example: Simple**
             ::
@@ -1234,7 +1250,9 @@ class PlotterBase(Plottable):
             return full_url
         elif (render is True) or in_ipython():
             from IPython.core.display import HTML
-            return HTML(make_iframe(full_url, self._height))
+            return HTML(make_iframe(full_url, self._height, extra_html=extra_html, override_html_style=override_html_style))
+        elif in_databricks():
+            return make_iframe(full_url, self._height, extra_html=extra_html, override_html_style=override_html_style)
         else:
             import webbrowser
             webbrowser.open(full_url)
@@ -1377,7 +1395,8 @@ class PlotterBase(Plottable):
                 or isinstance(graph, pa.Table) \
                 or ( not (maybe_cudf is None) and isinstance(graph, maybe_cudf.DataFrame) ) \
                 or ( not (maybe_dask_cudf is None) and isinstance(graph, maybe_dask_cudf.DataFrame) ) \
-                or ( not (maybe_dask_dataframe is None) and isinstance(graph, maybe_dask_dataframe.DataFrame) ):
+                or ( not (maybe_dask_dataframe is None) and isinstance(graph, maybe_dask_dataframe.DataFrame) ) \
+                or ( not (maybe_spark is None) and isinstance(graph, pyspark.sql.dataframe.DataFrame) ):
             return g._make_dataset(graph, nodes, name, description, mode, metadata, memoize)
 
         try:
@@ -1399,7 +1418,7 @@ class PlotterBase(Plottable):
         except ImportError:
             pass
 
-        error('Expected Pandas/Arrow/cuDF dataframe(s) or Igraph/NetworkX graph.')
+        error('Expected Pandas/Arrow/cuDF/Spark dataframe(s) or igraph/NetworkX graph.')
 
 
     # Sanitize node/edge dataframe by
@@ -1641,6 +1660,12 @@ class PlotterBase(Plottable):
             logger.debug('ddf->arrow via df hash check')
             ddf = table.persist()
             df = ddf.compute()
+            return self._table_to_arrow(df, memoize)
+
+        if not (maybe_spark is None) and isinstance(table, maybe_spark.sql.dataframe.DataFrame):
+            logger.debug('spark->arrow via df')
+            df = table.toPandas()
+            #TODO push the hash check to Spark
             return self._table_to_arrow(df, memoize)
 
         raise Exception('Unknown type %s: Could not convert data to Arrow' % str(type(table)))